Add files via upload
this is a function to show the feature maps of a cnn at once. any kind of addition is  welcome, be it documentation or generalizing the code or even adding new utilities function that will make coding a bit easier. this is just a small project that is done on a whim.